requirements.txt
PyGithub==2.1.1
requests==2.31.0 